Diagram - toolbars refactoring + tests
